+++ lib/router.dart
@@ -1,4 +1,5 @@
 import 'package:go_router/go_router.dart';
+import 'package:rakuten_books_viewer/page/search_page.dart';
 
 import 'page/home_page.dart';
 
@@ -9,6 +10,28 @@ final route = GoRouter(
       builder: (context, state) {
         return const HomePage();
       },
+      routes: [
+        GoRoute(
+          name: SearchPage.screenName,
+          path: SearchPage.routeName,
+          builder: (context, state) {
+            return const SearchPage();
+          },
+        )
+      ],
     ),
   ],
 );
+
+/*
+final route = GoRouter(
+  routes: [
+    GoRoute(
+      path: SearchPage.routeName,
+      builder: (context, state) {
+        return const SearchPage();
+      },
+    ),
+  ],
+);
+*/
